+++ jaziku/core/stations.py
@@ -77,6 +77,16 @@ def prepare_all_stations(stations_list, prepare_data_for_data_analysis, prepare_
             station.calculate_common_and_process_period()
         console.msg(_("done"), color='green')
 
+        # global common period
+        console.msg(_("Calculate, check and set the global process period ...... "), newline=False)
+        calculate_process_period(stations_list)
+        console.msg(_("done"), color='green')
+        # show the period calculated if is defined as maximum
+        if not env.config_run.settings['process_period']:
+            console.msg(_("   The maximum global common period is:"), newline=False)
+            console.msg("{0}-{1}".format(env.globals_vars.PROCESS_PERIOD['start'],
+                                          env.globals_vars.PROCESS_PERIOD['end']), color='cyan')
+
         # data, date and null
         console.msg(_("Prepare data, date and null in process period ........... "), newline=False)
         for station in stations_list:
@@ -84,6 +94,7 @@ def prepare_all_stations(stations_list, prepare_data_for_data_analysis, prepare_
             station.var_I.calculate_data_date_and_nulls_in_period()
         console.msg(_("done"), color='green')
 
+        # check the consistent data for all stations
         if env.config_run.settings['consistent_data'] is not False:
             console.msg(_("Check if the data are consistent for var_D and var_I:"))
             for station in stations_list:
@@ -133,7 +144,7 @@ def prepare_all_stations(stations_list, prepare_data_for_data_analysis, prepare_
 
 
 def global_process_period(stations_list):
-    """Calculate the global common period of all stations
+    """Calculate the maximum global common period of all stations
     based on all common process period of all series
 
     :arg:
@@ -141,16 +152,14 @@ def global_process_period(stations_list):
     :return:
         'start': start year of global process period
         'end': end year of global process period
-        'dates': chronological order list of all date (N-monthly or daily)
-                 inside of global process period
     """
     firsts = True
     for station in stations_list:
         if firsts:
-            global_common_date = set(station.var_D.date_in_process_period)
+            global_common_date = set(station.var_D.date) & set(station.var_I.date)
             firsts = False
         else:
-            global_common_date = global_common_date & set(station.var_D.date_in_process_period)
+            global_common_date = global_common_date & set(station.var_D.date) & set(station.var_I.date)
 
     global_common_date = list(global_common_date)
     global_common_date.sort()
@@ -159,16 +168,70 @@ def global_process_period(stations_list):
     if not global_common_date:
         return False
 
+    start_year = global_common_date[0].year + 1
+    end_year = global_common_date[-1].year - 1
+
+    return {'start': start_year,
+             'end': end_year}
+
+
+def calculate_process_period(stations_list):
+    """Check and set the process period, this is the
+    maximum global common period or the particular period
+    but this inside the maximum global common period
+
+    :arg:
+        stations: list of all stations
+    :return by reference:
+        env.globals_vars.PROCESS_PERIOD:
+            'start': start year of process period
+            'end': end year of process period
+    """
+
+    # calculate the maximum global common period
+    max_global_period = global_process_period(stations_list)
+
+    if max_global_period is False:
+        console.msg_error(_(
+                    "There is no a global common period for all stations,\n"
+                    "this mean that two or more stations don't have any\n"
+                    "data for a common date. Please check the data of\n"
+                    "stations."))
+
+    # if the period was defined in runfile
     if env.config_run.settings['process_period']:
-        # return the process period if this was defined inside the runfile
-        return {'start': env.config_run.settings['process_period']['start'],
-                 'end': env.config_run.settings['process_period']['end'],
-                 'start_date': date(env.config_run.settings['process_period']['start'], 1, 1),
-                 'end_date': date(env.config_run.settings['process_period']['end'], 12, 31 if env.var_D.is_daily() else 1),
-                 'dates': global_common_date}
+        # check if the period defined are inside the global period
+        if not (env.config_run.settings['process_period']['start'] >= max_global_period['start'] and
+                env.config_run.settings['process_period']['end'] <= max_global_period['end']):
+                console.msg_error(_(
+                    "The period defined in argument {0}-{1} is outside in the\n"
+                    "maximum global common period possible {2}-{3}.")
+                .format(env.config_run.settings['process_period']['start'],
+                    env.config_run.settings['process_period']['end'],
+                    max_global_period['start'], max_global_period['end']))
+
+        # set the process_period
+        env.globals_vars.PROCESS_PERIOD = \
+            {'start': env.config_run.settings['process_period']['start'],
+             'end': env.config_run.settings['process_period']['end']}
+
+        # check if the common period at least 3 years before calculate the process period
+        if (env.globals_vars.PROCESS_PERIOD['end'] - env.globals_vars.PROCESS_PERIOD['start']) < 3:
+            console.msg_error(_("The process period {0}-{1} has less than 3 years,\n"
+                                "Jaziku need at least 3 years for run.")
+                .format(env.globals_vars.PROCESS_PERIOD['start'], env.globals_vars.PROCESS_PERIOD['end']))
+
+    # if the period is the maximum global period (maximum in runfile)
     else:
-        return {'start': global_common_date[0].year,
-                 'end': global_common_date[-1].year,
-                 'start_date': global_common_date[0],
-                 'end_date': global_common_date[-1],
-                 'dates': global_common_date}
+        # set the process_period
+        env.globals_vars.PROCESS_PERIOD = \
+            {'start': max_global_period['start'],
+             'end': max_global_period['end']}
+
+        # check if the common period at least 3 years before calculate the process period
+        if (env.globals_vars.PROCESS_PERIOD['end'] - env.globals_vars.PROCESS_PERIOD['start']) < 3:
+            console.msg_error(_("The maximum global common period possible of intersection\n"
+                                "for all stations and {0} time series is {1}-{2},\n"
+                                "Jaziku need at least 3 years for run, please check the\n"
+                                "data of stations.")
+                .format(env.var_I.TYPE_SERIES, env.globals_vars.PROCESS_PERIOD['start'], env.globals_vars.PROCESS_PERIOD['end']))